+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Released
 
+## [0.2.0] - 2024-02-23
+
+- added options to specify APNs and Android sound
+- default sound for Android notifications
+- ruby update to 3.3
+
 ## [0.1.1] - 2023-12-20
 
 - Initial release